embryo-quality-prediction/requirements.txt
# Core Data Science and Machine Learning Libraries
numpy==1.24.3         # Numerical computing library for array operations and linear algebra
pandas==2.0.3         # Data manipulation and analysis library
matplotlib==3.7.2     # Comprehensive plotting and visualization library
seaborn==0.12.2       # Statistical data visualization based on matplotlib
scikit-learn==1.3.0   # Machine learning algorithms for classification, regression, clustering, etc.

# Deep Learning Libraries
torch==2.0.1          # PyTorch deep learning framework
torchvision==0.15.2   # Computer vision utilities and datasets for PyTorch

# Image Processing
Pillow==9.5.0         # Python Imaging Library for opening, manipulating, and saving images
opencv-python==4.8.0.74  # Computer vision library with image processing functions

# Progress Tracking and Utilities
tqdm==4.65.0          # Progress bar for loops and data processing

# Web Application Framework (for Evaluation Dashboard)
Flask==2.3.2          # Lightweight web application framework
Werkzeug==2.3.6       # WSGI utility library for Flask
Jinja2==3.1.2         # Template engine for Flask
# Use --ignore-installed for problematic packages
itsdangerous==2.1.2   # Securely sign data for Flask
click==8.1.3          # Command-line interface creation toolkit
markdown==3.4.3       # Markdown to HTML converter for documentation rendering
bleach==6.0.0         # HTML sanitizing library for secure content rendering
python-markdown-math==0.8  # Math extension for Python-Markdown
Pygments==2.15.1      # Syntax highlighting for code blocks in documentation
Flask-WTF==1.1.1      # Flask integration with WTForms for form handling

# Data Visualization for Dashboard
plotly==5.15.0        # Interactive visualization library
chart.js==3.9.1       # JavaScript charting library for data visualization
# Note: Chart.js is loaded via CDN in the HTML templates, not installed via pip

# Model Evaluation and Metrics
scipy==1.10.1         # Scientific computing library with statistical functions

# Database
mysql-connector-python==8.0.33  # MySQL database connector for Python

# Utilities
python-dotenv==1.0.0  # Environment variable management
requests==2.31.0      # HTTP library for API requests
watchdog==3.0.0       # File system events monitoring for development
gunicorn==21.2.0      # WSGI HTTP server for production deployment

# Development Tools - commented out for Colab compatibility
# ipython==8.14.0      # Enhanced interactive Python shell
# jupyter==1.0.0       # Interactive computing environment

# Documentation - commented out for Colab compatibility
# Sphinx==7.1.2        # Documentation generator

# Testing - commented out for Colab compatibility
# pytest==7.4.0        # Testing framework

# Colab-specific requirements
# These are pre-installed in Colab but listed here for reference
# google-colab
# google-auth
# pydrive

# Lion optimizer
lion-pytorch==0.1.2   # Lion optimizer implementation for PyTorch